docs/distros.md: cleaner nixpkgs overlay
I believe that using `fetchurl` is very cumbersome, since `fetchFromGitHub` already exists and the source is already a GitHub repo. It also causes weird problems that I can't quite figure out when you try to update to a later commit. By using `fetchFromGitHub` instead, updating just means you need to update the rev and hash. This also has the added benefit of not needing to override `unpackPhase` at all.

I've also decided to move the WeirdTreeThing repo sources off to a different variable `wttsrc` instead of trying to override `src` or create a new variable `srcs`. This makes the intent of the overlay clearer and lets the alsa-ucm-conf tarball defined in `src` update independently of the overlay.